Finalize authentication flow and notes access
- Implement user registration with validation and auto-login
- Configure Spring Security with form login and protected routes
- Redirect authenticated users to notes page
- Ensure logout redirects to public home page
- Clean up controllers and remove ambiguous route mappings

+++ lab10/src/main/java/com/example/lab10/model/User.java
@@ -9,7 +9,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(columnDefinition = "INTEGER")
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false, unique = true)
     private String username;
@@ -17,18 +17,17 @@ public class User {
     @Column(nullable = false, unique = true)
     private String email;
 
-    // IMPORTANT: no validation annotations here
     @Column(nullable = false)
     private String password;
 
     public User() {}
 
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
     public String getUsername() { return username; }
     public String getEmail() { return email; }
     public String getPassword() { return password; }
 
-    public void setId(Long id) { this.id = id; }
+    public void setId(Integer id) { this.id = id; }
     public void setUsername(String username) { this.username = username; }
     public void setEmail(String email) { this.email = email; }
     public void setPassword(String password) { this.password = password; }
